# Source/Training/modules/studentobject-http-options/studentobject-http-options-impl/rest-openapi.yaml
info:
    description: "StudentobjectHttpOptions REST API"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "StudentobjectHttpOptions"
    version: v1.0
openapi: 3.0.1
components:
  schemas:
    StudentsObject:
      properties:
        studentId:
          type: integer
          format: int32
        studentName:
          type: string
        departmentId:
          type: integer
          format: int32
paths:
...
  /update-student:
    post:
      operationId: updateStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentsObject"
          application/xml:
            schema:
              $ref: "#/components/schemas/StudentsObject"
      responses:
        200:
          description: "Success Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentsObject"
            application/xml:
              schema:
                $ref: "#/components/schemas/StudentsObject"
      tags: ["StudentsObject"]
...